+++ modules/localplay/xbmc/xbmc.controller.php
@@ -229,8 +229,8 @@ class AmpacheXbmc extends localplay_controller
 
         $user_id = $user_id ? $user_id : Core::get_global('user')->id;
 
-        Preference::update('xbmc_active', $user_id, (int) ($uid));
-        AmpConfig::set('xbmc_active', (int) ($uid), true);
+        Preference::update('xbmc_active', $user_id, $uid);
+        AmpConfig::set('xbmc_active', $uid, true);
 
         return true;
     } // set_active_instance
